fix(shortcuts): register global shortcuts at launch
MenuBarExtra content with .menuBarExtraStyle(.window) is built lazily, so
the .task registering KeyboardShortcuts handlers never ran until the user
first opened the popover. Move registration into
AppDelegate.applicationDidFinishLaunching.

Fixes #199
